Refactor: Add <Home/> and refractor routing
- Combined <AddItem/> and <AddItemFirebase/> components
- Add a <Home /> component
- Add a <Home /> component route with <List /> component new route
- Add wireframe images
