+++ src/components/Home/Home.component.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Home = () => {
+  return <div> Home Page</div>;
+};
+
+export default Home;
