+++ src/components/component.index.js
@@ -1,5 +1,6 @@
+import Home from './Home/Home.component.jsx';
 import AddItem from './AddItem/AddItem.component';
 import List from './List/List.component';
 import Footer from './Footer/Footer.component.jsx';
 
-export { AddItem, List, Footer };
+export { Home, AddItem, List, Footer };
